Set DotNetUseShippingVersions. (#595)
This disables the `42.42.42.42` assembly versions when building locally. Instead, a date based version is used with a `-0` suffix.

This allows for locally built assemblies to replace official built assemblies in an install folder. This is useful for debugging (can replace a release binary with a debug binary), and for quick prototyping without having to rebuild the whole dotnet-try package and uninstall-reinstall.